requirements.txt
click==8.1.3
colorama==0.4.5
cycler==0.11.0
Flask==2.1.2
fonttools==4.34.4
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
scipy==1.8.1
six==1.16.0
Werkzeug==2.1.2
